requirements.txt
coremltools==5.0b3
gunicorn==20.1.0
Flask==1.1.2
Pillow==7.2.0